feat: gitpt is now a CLI wizard!
It can now generate commit messages of any length you desire and even lets you choose whether to include that fancy prefix before the message.  We've also updated the README with all the juicy details about how to install and use gitpt. So go forth and create commits with style!